--- lib/home.dart
@@ -1,45 +1,0 @@
-import 'package:flutter/material.dart';
-
-class HomePage extends StatefulWidget {
-  @override
-  _HomePageState createState() => _HomePageState();
-}
-
-class _HomePageState extends State<HomePage> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-        body: Container(
-          color:  Colors.cyan[600],
-          child: ListView(
-            children:[
-              Padding(
-                padding: EdgeInsets.all(30),
-                child: Text(
-                  'Watch List',
-                  style: TextStyle(
-                      fontFamily: 'Montserrat',
-                      color: Colors.white,
-                      fontWeight: FontWeight.bold,
-                      fontSize: 25
-                  ),
-                ),
-              ),
-              Container(
-                height: MediaQuery.of(context).size.height - 185.0,
-                decoration: BoxDecoration(
-                  color: Colors.white,
-                  borderRadius: BorderRadius.only(topLeft: Radius.circular(75.0)),
-                ),
-                child: ListView(
-                  children: <Widget>[
-
-                  ],
-                ),
-              )
-            ],
-          ),
-        )
-    );
-  }
-}
