+++ lib/watchlist.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+import 'main.dart';
+
+class WatchListPage extends StatefulWidget {
+  @override
+  _WatchListPageState createState() => _WatchListPageState();
+}
+
+class _WatchListPageState extends State<WatchListPage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+        body: PageTheme().pageTheme('Watch List', ListView(),context)
+    );
+  }
+}
